fixedstr

Type Alias str8

pub type str8 = tstr<8>;

Types for small strings that use an efficient representation underneath. Alias for internal type [tstr]<8>. A str8 is 8 bytes and can hold string of up to 7 bytes. See documentation for the aliased [tstr] type.

Example:

 let mut s = str8::from("aλc");
 assert_eq!(s.capacity(),7);
 assert_eq!(s.push("1234567"), "4567");
 assert_eq!(s,"aλc123");
 assert_eq!(s.charlen(), 6);
 assert_eq!(s.len(), 7);  

Aliased Type

struct str8 { /* private fields */ }